+++ SpeedoPoint/hotellisting.h
@@ -29,7 +29,7 @@ public:
     hotellisting(hotel hot, country location, int price, int index, int area_, bool wifi_, bool breakfast, int numPeople, bool refund, bool dinner_, bool pet, std::string roomtype);
     reservation* reserve(user* acc, date d, int days, int adults, int children);
     int getPricePerNight();
-    int getHotelRating();
+    float getHotelRating();
     hotel getHotel();
     country getLoc();
     int getCityIndex();
